requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
Babel==2.15.0
boto3==1.21.38
botocore==1.24.38
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
dataclasses==0.6
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.8.1
django-ratelimit==3.0.1
django-storages==1.12.3
djangorestframework==3.15.1
dnspython==2.7.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.0
mongoengine==0.29.1
ollama==0.2.1
Pillow==9.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.3
py-moneyed==3.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==3.11.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.1
reportlab==4.2.0
requests==2.32.2
s3transfer==0.5.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.2
typing_extensions==4.11.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
